remove float cast

+++ src/android/IonicKeyboard.java
@@ -18,7 +18,7 @@ public class IonicKeyboard extends CordovaPlugin{
         //http://developer.android.com/guide/practices/screens_support.html
         DisplayMetrics dm = new DisplayMetrics();
         cordova.getActivity().getWindowManager().getDefaultDisplay().getMetrics(dm);
-        final float density = (float)(dm.density);
+        final float density = dm.density;
         
         final CordovaWebView appView = webView;
